A lot a say a lot of change and features

+++ src/App/Admin/Company/Role/Controllers/CompanyRoleDestroyController.php
@@ -3,10 +3,12 @@
     namespace App\Admin\Company\Role\Controllers;
 
     use Domain\Company\Models\Company;
+    use Domain\Role\Mappings\RoleTableMap;
+    use Domain\Role\Models\Role;
     use Illuminate\Http\JsonResponse;
-    use Spatie\Permission\Models\Role;
     use Support\Helpers\Response\Popups\Notifications\SimpleNotification;
     use Support\Helpers\Response\Response;
+    use Symfony\Component\HttpFoundation\Response as ResponseCodes;
 
     class CompanyRoleDestroyController
     {
@@ -18,10 +20,19 @@
          */
         public function action(Company $company, Role $role): JsonResponse
         {
-            $role->delete();
+            $response = new Response();
 
-            return Response::create()
-                ->addPopup(new SimpleNotification(trans('response.success.delete.company.role')))
+            if ($role->name !== RoleTableMap::MAIN_ROLE) {
+                $role->delete();
+
+                return $response->addPopup(new SimpleNotification(trans('response.success.delete.company.role')))
+                    ->toJson();
+            }
+
+            return $response->addPopup(
+                new SimpleNotification(trans('response.error.delete.admin.role')),
+                ResponseCodes::HTTP_BAD_REQUEST
+            )
                 ->toJson();
         }
     }
